Use hostname in disk resource names
Otherwise "disk resources" view in UpCloud control panel
looks super confusing, as all root disks are "terraform-os-disk-0".
If multiple disks are detached from hosts then it's hard to
say which disks belong to which instances. Disk UUIDs can be
used to match disks to hosts in case of an emergency though.